requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coverage==6.3.2
ecdsa==0.17.0
et-xmlfile==1.1.0
fastapi==0.74.1
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
iniconfig==1.1.1
joblib==1.1.0
Mako==1.1.6
MarkupSafe==2.1.0
numpy==1.22.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
threadpoolctl==3.1.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
zipp==3.7.0
